feat: implement commission on repay and refactor related code
- Implemented commission on repay feature
- Completed all required tasks
- Removed unused variables
- Renamed for clarity
- Cleaned up code

+++ contracts/pricefeeds/ConstantPriceFeed.sol
@@ -10,7 +10,7 @@ import "../interfaces/IPriceFeed.sol";
  */
 contract ConstantPriceFeed is IPriceFeed {
     /// @notice Version of the price feed
-    uint public constant override version = 1;
+    uint256 public constant override version = 1;
 
     /// @notice Description of the price feed
     string public constant description = "Constant price feed";
